Sync pif.device when its type is vlan or sriov
For pif which is vlan-master or sriov-logic-pif, its pif.device
comes from its vlan-slave or sriov-physical-pif pif.device.
After RPU, the network interface name may change. Its corresponding
physical pif.device is synced already. The vlan or sriov
non-physical pif.device should be synced accordingly.
For non-physical and unattached pifs (like vlan, sriov pifs that
are not used), the pif.device need to be updated too. So,
refresh_all get all pifs on the host and in refresh_internal the
pif fields physical and currently-attached are carefully distinguished
to keep the original logic not changed.

Signed-off-by: Changlei Li <changlei.li@cloud.com>