CP-308798: Dynamic control NBD firewalld service
For iptables, NBD is a special case as it uses a separate script,
`nbd-firewall-config.sh`, to enable the port on each requested interface.
Therefore, `update_firewall_status` needs a new argument, `interfaces`,
which specifies the list of interfaces on which the port should be enabled.

For firewalld, NBD follows the same procedure as other services. In this
case, simply log the value of the `interfaces` argument to avoid `make
check` warnings.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>