CP-309535: Add a service to update firewalld
When the firewalld restarts, all firewalld services are reset to the
default status. This function should be called to update these firewalld
services to the correct status. Xapi will expose an xe command line for
this scenario.
Add a systemd service update-xapi-firewalld.service to update the status of
all the firewalld services to match the state of the corresponding
services when the firewalld restarts.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>