xapi_pbd: use HA shared SR constraint violation when plugging and unplugging
When unplugging a pbd, enabling a host, or adding a vbd, the shared SR
constraint violation could be violated, but the error used in these cases
was that the operation blocked the failover planning. This was confusing
because the main reason was not mentioned in the error. Instead use the
SR constraint violation error, and log a more descriptive message in the
logs as info, because these can happen during normal operation and
there's nothing dodgy going on.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>