http-lib: Add filename hint to file response
When serving files, a filename can be hinted to the client, so they can
download the with the correct filename instead of showing the contents
on the screen. For example, with `curl -JO`.

Don't change behaviour unless explicitly requested, users like the xapi
fileserver serve html files which are not meant to be downloaded.

See https://httpwg.org/specs/rfc6266.html#rfc.section.4.2

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>