doc: update better VM revert proposal
This proposal is 11 years old and the architecture of xapi ahs changed
enough that the proposal needs changing.

In particular, SMAPIv3 has been introduced which means that there's an
xapi storage interface shared between the two SMAPI backend, and that
the fallback the logic that was done at SMAPI level now must be done at the
level of xapi.

I've also elected to further restrict the fallback behaviour to VM
reverts, VDI revert will require backend support: it reduces the chances of
failures during revert resulting in wrong fields in the database.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>