CA-417951: Split checks for migration and detecting RPU
Turns out that the previous fixes on these were too strict. As the
checks for allowing migration and detecting RPU are a bit different now.

For migration we want to make sure we are migrating on the same or
higher versions for: platform, xapi, and xen versions.

For detecting RPU we want the pool to have the same versions for
platform and xapi build (only major and minor numbers).

Signed-off-by: Gabriel Buica <danutgabriel.buica@citrix.com>