CA-417951: Host.create now takes a software version argument
Currently, when a new host is created on the coordinator, it has only a
minimal record added for it in the database. And this record only adds
the platform version. While the rest is populated when `db_sync` is
called.

With the new checks for detecting RPU, this results in a 30s interval
when the coordinator thinks an RPU is in progress when a new host is
added to a pool.

Because we already check for for xapi build version, major and minor
numbers, in the pre-join checks, it is safe to add this in the minimal
record created for a new host (before db_sync gets called).

Signed-off-by: Gabriel Buica <danutgabriel.buica@citrix.com>