xapi_vm_snapshot: move VDI-related DB operations inside revert_vbds
This moves them outside of the try catch that destroys newly-cloned
disks, but this is safe since these don't change the state of the host
like the destroy vifs.

This allows to simplify the type needed to be able to clean up the new
VDIs in case of failure.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>
(cherry picked from commit 37f15f88d22850418a32d170ccedc1cbf952cd5f)