Extend diagnostic-timing-stats to optionally show counts
The Host.get_diagnostic_timing_stats API call shows mean and SD of time
spent. We would like to know how many samples were used for this but
can't simply add this because it could break existing clients. Add an
optional flag to include them with a default that maintains the existing
behaviour.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>