localdb: expose option-based get functions
Allows callers to avoid exceptions, the previous get is now called
get_exn so it's clear which users have to be have to be mindful of
exceptions.

Not all get_exn were converted to get, previous behaviour was
widespread, and doing the change without changing behaviour is not
trivial, better to do it only once its detected.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>