Move xenopsd's xenctrl_ext to ocaml/libs/
xenctrl_ext contains C bindings to xenctrl and Xen. This is a place to
add C bindings that maybe are not yet evailable in Xen.

Currently their visibility is limited to Xenopsd. This patch moves them
into their own library - and hence makes them accessible from Xapi and
other code in this repository.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>