xenopsd/xc: adapt claim_pages to new single numa node version (CP-53658)
Now the numa node needs to be passed. A special value of (~0U) is used to
signify that no node is meant to be used. Since this is arch-dependent,
and contained in a long in x86_64, an int is used to encode the value.

Also remove the exception that was guarding the codepath to use this case

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>