xapi-consts: use a variant to reify valid values of cluster stack
Until now, the valid string values were not written anywhere, change the
situation.

Ideally this would be done using an enum in the idl, but unfortunately
this changes types of existing parameters in API calls, so it's quite
risky. Instead have a conservative approach to only enumerate the valid
values and make Constants the only source of truth for these values,
including default ones, which were scatterred around previously.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>