CP-309791: Update samba to 4.21
Microsoft Windows is doing security hardening
https://bugzilla.samba.org/show_bug.cgi?id=15876
Our samba needs to update accordingly.

There are some interface and configuration updates, thus xapi
needs to be updated as well. This update is intended te be
compatible with old and new samba, to decouple the merge of samba
update. The old interface support will be dropped after samba merge.

Signed-off-by: Lin Liu <Lin.Liu01@cloud.com>