XSI-2025: Only abort HA host reenable for unpluggable PIFs
Currently, abort_if_network_attached_to_protected_vms is used to abort
unplugs and host reenables under HA when there are any unplugged PIFs.
This isn't necessary for pluggable PIFs as it is a quick local operation
to do so. The only time we should abort is for bond slave PIFs as these
are not pluggable and should be blocked for VM agility.

Signed-off-by: Steven Woods <steven.woods@cloud.com>