CA-418227: Handle exceptions in At_least_once_more.again
At_least_once_more is used by `consider_enabling_host_request`, which is an
important function to get hosts marked as `enabled`. The purpose of this
is to prevent running `consider_enabling_host` many times unnecessarily.

The problem with the `again` function is that it does not handle any
exception raised by the operation `f`. If an exception occurs, then the
thread exists without setting `in_progress` to `false`, which prevents
the operation from ever getting executied again.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>