CP-310425: Add username in VNC console error messages
When a VNC console connection is rejected due to the session limit,
users want to know which user(s) are currently connected. However,
displaying usernames in HTTP error responses may raise privacy concerns
in some deployments, so also add a configure to enable/disable the
display of the usernames.

The main changes are:
1. To contain the active users in the response message,
changed the active_connections to record the existing users
and use the unique session id to identify them in case
multiple connections have the same user name.

2. Use Http_svr.escape to escape html special characters

3. Added `include_console_username_in_error` to enable/disable
   the display of user names

Signed-off-by: Stephen Cheng <stephen.cheng@citrix.com>