CA-420968: compute the amount of physical cores available on a NUMA node set
Could also compute it by multiplying it with [threads_per_core],
but I'm not sure how that'd interact with [smt=false] in Xen.
Also to future-proof this I wouldn't want to rely on an entirely
symmetrical architecture
(although it'd be very rare to have anything other than 2 on x86-64,
 or to have hyperthreading on in one socket, and off in another).

Note that core ids are not unique (there is a core `0` on both socket 0 and
socket 1 for example), so only work with number of cores in the topology code.

Could've created a CoreSocketSet instead (assuming that no higher grouping than
sockets would exist in the future), but for now don't make too many assumptions
about topology.

No functional change.

Signed-off-by: Edwin Török <edwin.torok@citrix.com>