CA-420968: introduce an explicit name for the current NUMA policy: Prio_mem_only
The current NUMA policy prioritizes reducing cross-NUMA node memory traffic by
picking the smallest set of NUMA nodes that fit a VM.
It doesn't look at how this affects CPU overload within a NUMA node, or whether
the local bandwidth of each NUMA node is balanced or not.

Give this policy an explicit name, `Prio_mem_only`, and when the "compat" setting
in `xenopsd.conf` is used (`numa-placement=true`), then explicitly use this
policy instead of Best-effort.

Currently Best-effort is still equivalent to this policy, but that'll change in
a follow-up commit.
Introduce a new xenopsd.conf entry `numa-best-effort-prio-mem-only`,
which can be used to explicitly revert best effort to the current policy.
(currently this is a no-op, because there is only one best-effort policy).

Future policies should also look at CPU overload.

No functional change.

Signed-off-by: Edwin Török <edwin.torok@citrix.com>