CA-420968: ensure compatibility between NUMARequest.fits and plan
The planner explicitly looks at the NUMARequest fields and checks that they are
non-zero.
However if more fields get added in the future this leads to an assertion
failure, where the planner thinks it has found a solution, but NUMARequest.fits
returns false.

Ensure consistency: use `fits` in the planner to check that we've reached a
solution. If the remaining request doesn't fit into an empty node, then the
request is not empty.

Signed-off-by: Edwin Török <edwin.torok@citrix.com>