s/xe-syslog-reconfigure: Keep disclaimer in remote.conf
The behaviour is same, except that the file will remain (with empty
rules) even if the forwarding log has been disabled (and then prevent
the disclaimer loss).

Wording slightly modified to reduce the risk of concurrent management
of this file, if the file is still edited by other mean, it will be
done knowingly (that x-s-r may override any change soon or later).

For the record remote.conf was introduced in:
468eb75dddfea6db512a8bfb4860ff2042efab66 .

I am assuming that the presence of file is not checked elsewhere that
in xen-api (currently only s/x-s-r is referencing this file).

Signed-off-by: Philippe Coval <philippe.coval@vates.tech>