ocaml: reduce and guard all users of Pool.get_all |> List.hd
During the bringup of a new platform for XCP-ng, one of the users has
raised an exception because the database doesn't have any pool record.

This change makes sure that the there's more information on the
exception.

Also changes other places where Pool.get_all to make it obvious that the
result is matched and List.hd is less likely to be used.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>