increase max supported NVMe request size
The current default value for the NVMe MDTS parameter exposed in QEMU emulated
NMVe devices is 7 (max 512KiB requests).  However there seems to be an
internal Windows Server 2025 issue that possibly triggers when splitting
bigger requests into smaller on in the NVMe Windows driver.

Increase the exposed MDTS value on the emulated QEMU NVMe device to 9 (max 2MiB
request size), as that seems to drop the reproduction rate of the issue.

Discussion is ongoing with Microsoft to get the issue identified and
possibly sorted on their end.  For the time being apply this mitigation in
qemu-wrapper as a workaround.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>