rate-limit: Handle rate limited requests in FIFO queue
The current implementation of rate limiting had severe fairness issues.
These have been resolved through the addition of a request queue, to
which rate limited requests are added. A worker thread sleeps until its
associated token bucket has enough tokens to handle the request at the
head of the queue, calls it, and sleeps until the next request is ready.

Signed-off-by: Christian Pardillo Laursen <christian.pardillolaursen@citrix.com>