CA-421914: preserve Host.numa_affinity_policy across pool join
When `numa-placement` was true (the default in XS9),
configuring a host's `numa-affinity-policy` to `any` got lost during pool join.

This is because `create_params` is maintained by hand and needs to be extended
every time we add a new field to the datamodel.

Update `create_params` to propagate `numa_affinity_policy` on pool join.

Before pool join:
```
[root@genuk-21-09d ~]# xe host-list params=all|grep numa
                  numa-affinity-policy ( RW): any
```

After pool join:
```
[root@genuk-21-09d ~]# xe host-list params=all|grep numa
                  numa-affinity-policy ( RW): any
                  numa-affinity-policy ( RW): any
```

(prior to this bugfix the newly joined host would revert to default_policy).

The correct policy is also set in xenopsd:
```
2025-12-15T15:12:09.029478+00:00 genuk-21-09d xenopsd-xc: [ info||134
|host.set_numa_affinity_policy R:99aecf4c494f|xenops_server] Enforcing 'any'
NUMA affinity policy
```

Signed-off-by: Edwin Török <edwin.torok@citrix.com>