CA-422072 Add new fields to host create params
For the new added unit test which check host create
params, need to add the new fields to create params.
This check is mainly for pool join case. Without the
params, the values will be set to default values
when the supporter host obj is created. Then the value
may be lost after pool-join.
However, for the fields in the feature, the values will
be set by dbsync. So the values can be set correctly
after restart during the pool-join. So there will not be
defects.
On the other hand, there is also no harm for adding them
to create_params. It can pass the new added unit test
and set the value correctly when host object is created
during pool-join. So add them in this commit.

Signed-off-by: Changlei Li <changlei.li@cloud.com>