CA-422187,CA-422188: either always use claims or never use claims
Do not mix using claims with not using claims.
Xen cannot currently guarantee that it'll honour a VM's memory claim,
unless all other VMs also use claims.

Global claims have existed since a long time in Xen,
so this should be safe to do on both XS8 and XS9.

Safer defaults for global claims:

Xen may have already allocated some memory for the domain, and the overhead is
only an estimate.
A global claim failing is a hard failure, so instead use a more conservative
estimate: `memory.build_start_mib`.
This is similar to `required_host_free_mib`, but doesn't take overhead into
account.

Eventually we'd want to have another argument to the create hypercall that
tells it what NUMA node(s) to use, and then we can include all the overhead too
there.

For the single node claim keep the amount as it was, it is only a best effort
claim.

Do not claim shadow_mib, it has already been allocated:

When rebooting lots of VMs in parallel we might run out of memory
and fail to boot all the VMs again.
This is because we overestimate the amount of memory required, and claim too
much. That memory is released when the domain build finishes, but when building
domains in parallel it'll temporarily result in an out of memory error.

Instead try to claim only what is left to be allocated: the p2m map and shadow
map have already been allocated by this point, i.e. claim just the bare
minimum.

Signed-off-by: Edwin Török <edwin.torok@citrix.com>