coverage: remove dependency on bisect_ppx
This infrastructure has been unused for many years (I think it only ever
worked for some months during 2019), and it's making ocaml 5.4 difficult
to adopt, so drop it.

I'm leaving the instrumentation in the dune metadata because it's
disabled by default, and the document about coverage, because it's a
historical document.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>