CA-422187: fix NUMA on XS8
On XS8 we always raise an exception when attempting to claim from a single
node.
We wanted to only use soft affinity when the single node claim succeeded (which
is the correct fix on XS9, where this API is available).
However this meant that we've effectively completely disabled NUMA support on
XS8, without any way to turn it on.

Always use soft affinity when the single-node claim API is unavailable, this
should keep NUMA working on XS8.

On XS9 Xen itself would never raise ENOSYS (it has a `err = errno = 0` on
ENOSYS).

Fixes: fb66dfc03 ("CA-421847: set vcpu affinity if node claim succeeded")

Signed-off-by: Edwin Török <edwin.torok@citrix.com>