CA-422187: safer defaults for global claims
Xen may have already allocated some memory for the domain, and the overhead is
only an estimate.
A global claim failing is a hard failure, so instead use a more conservative
estimate: `memory.build_start_mib`.
This is similar to `required_host_free_mib`, but doesn't take overhead into
account.

Eventually we'd want to have another argument to the create hypercall that
tells it what NUMA node(s) to use, and then we can include all the overhead too
there.

For the single node claim keep the amount as it was, it is only a best effort
claim.

Fixes: 060d79250 ("CA-422188: either always use claims or never use claims")

Signed-off-by: Edwin Török <edwin.torok@citrix.com>