CA-422187: do not claim shadow_mib, it has already been allocated
When rebooting lots of VMs in parallel we might run out of memory
and fail to boot all the VMs again.
This is because we overestimate the amount of memory required, and claim too
much. That memory is released when the domain build finishes, but when building
domains in parallel it'll temporarily result in an out of memory error.

Instead try to claim only what is left to be allocated: the p2m map and shadow
map have already been allocated by this point.

Fixes: 95367e1aa ("CA-422187: safer defaults for global claims")

Signed-off-by: Edwin Török <edwin.torok@citrix.com>