xapi: Add new message-destroy-all API call
We've seen issues where a client will try to dismiss all their alerts
and fail due to the message-destroy call hitting the maximum XAPI
request size at ~5000 messages.

A new message-destroy-all API avoids this issue while saving the
bandwidth required to send a ref for each message. Users of the new
API call can choose to dismiss messages that arrived before or after
a certain date, or only those with a given priority.

Signed-off-by: Christian Pardillo Laursen <christian.pardillolaursen@citrix.com>