libs: add Stat module in Unixext to handle special device IDs
These device IDs are handled in an ad-hoc way, and none of the
implementations follow the actual linux behaviour.

The glibc-provided macros are used for the implementation, with a
pure-ocaml implementation that was useful to compare against while
implementing, which has been kept in the tests, to detect any
behavioural changes.

Because Unix.stat returns an `int` instead of an int64, the code does
not support all possible values for major supported by glibc. This
shouldn't be an issue in Linux since the device value is 32-bit wide.

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>